Address CodeRabbit review: unsigned CAN ticks, WDT log, sleep-lock docs
- Make lastCANChassisTick/lastCANHaldexTick and canServiceRxFault's
  everAliveTick uint32_t. As signed long they went negative after the
  32-bit millis() rollover (~24.8 days uptime), which would disable the
  everAliveTick > 0 guard that arms the drive-critical CAN restart.
- Log when the updateTriggers task-WDT subscription fails permanently
  instead of leaving the deadlock backstop silently absent.
- Document that the ESP_PM_NO_LIGHT_SLEEP toggle is currently redundant
  (light sleep is disabled globally in main.cpp) and kept as
  defense-in-depth if light sleep is ever re-enabled.

+++ include/OpenHaldexC6_defs.h
@@ -415,8 +415,12 @@ extern uint8_t udsBlockagePct;   // unconfirmed DID — always 0
 
 extern uint32_t alerts_to_enable;
 
-extern long lastCANChassisTick;
-extern long lastCANHaldexTick;
+// millis() timestamp of the last frame on each bus. Unsigned to match millis()
+// so the value never goes negative after the signed 32-bit rollover (~24.8 days
+// uptime) - a negative tick would silently defeat the everAliveTick > 0 guard
+// that arms the drive-critical CAN restart in canServiceRxFault.
+extern uint32_t lastCANChassisTick;
+extern uint32_t lastCANHaldexTick;
 extern volatile uint32_t lpChassisFrameCount; // chassis CAN frame counter for probe window activity check
 extern volatile uint32_t lpHaldexFrameCount;  // haldex CAN frame counter for standalone probe window activity check
 extern uint32_t canHealthTimeoutMs;
